add support for new scheduler/worker model introduced in 18.7
The new scheduler in 18.7 is now decoupled from the source code for registering
worker jobs. This means you need to explicitly send the scheduler a message
when the plugin is loaded indicating intent to enqueue periodic or calendar
based jobs. Likewise the worker needs to explicitly add the plugin job via the
Router module. Job names must be globally unique and match the name given to the
scheduler.